refactor(crawl_cpt_fx): rename 'crawler' to 'extractor'
