+++ plugins/src/fx_to_sap/extractor.py
